feat: Add utility function for merging CSS class names
This commit adds a new utility function `cn` to the `utils.ts` file. The `cn` function takes in multiple CSS class names as arguments and uses the `clsx` and `twMerge` functions to merge them into a single class name. This utility function will be useful for combining Tailwind CSS classes in the project.

Note: This commit message follows the conventional format of starting with a type (feat) followed by a brief description of the changes made.