feat: Add axios and json-server dependencies for API integration
This commit adds the "axios" and "json-server" dependencies to the package.json file. The "axios" library will be used for making HTTP requests to the API, while "json-server" will be used as a mock server for development and testing purposes. These dependencies enable API integration and facilitate the retrieval of vocabulary data for the MultipleChoice component in the Turkish practice page.

Note: This commit message follows the conventional format of starting with a type (feat) followed by a brief description of the changes made.